config: comment out options that we want to override in the makefile
If the config file specifies y or n for an option we are unable to
change it in the makefile, so comment out the relevant options for
crypto so they can be set via makefile.

There's an issue with DPDK and build order when we use makefile
cmd line options like we do to build the PMDs and APIs for
crypto and compression.  Basically DPDK will not compile if we
specify all of them on the command line.  Turns out if we specify
just the APIs in the command line and everything else in the
makefile we can build both crypto and compress just fine. This
allows us to avoid making folks install things like isa-l and
ipsec unless they actually are going to use compress or crypto.

Everyone, however, will build the 2 API modules for DPDK all of
the time but this should not be an issue.

Signed-off-by: paul luse <paul.e.luse@intel.com>
Signed-off-by: Jim Harris <james.r.harris@intel.com>
Signed-off-by: Dariusz Stojaczyk <dariuszx.stojaczyk@intel.com>
Signed-off-by: Vitaliy Mysak <vitaliy.mysak@intel.com>